% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_element.R
\name{element_path}
\alias{element_path}
\title{Theme Element for Image Grobs}
\usage{
element_path(
  alpha = NULL,
  colour = NA,
  hjust = NULL,
  vjust = NULL,
  color = NULL,
  angle = NULL,
  size = 0.5
)
}
\arguments{
\item{alpha}{The alpha channel, i.e. transparency level, as a numerical value
between 0 and 1.}

\item{colour, color}{The image will be colorized with this color. Use the
special character \code{"b/w"} to set it to black and white. For more information
on valid color names in ggplot2 see
\url{https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill}.}

\item{hjust, vjust}{The horizontal and vertical adjustment respectively.
Must be a numerical value between 0 and 1.}

\item{angle}{The angle of the element as a numerical value between 0° and 360°.}

\item{size}{The output grob size in \code{cm} (!).}
}
\value{
An S3 object of class \code{element}.
}
\description{
In conjunction with the \link[ggplot2:theme]{ggplot2::theme} system, the function \code{element_path()}
enables images in non-data components of the plot, e.g. axis text. It draws
images from valid image URLs, raster objects, or bitmap arrays.
}
\examples{
\donttest{
library(ggplot2)

df <- dplyr::mutate(mtcars,
  team = sample(c("LAC", "BUF", "DAL", "ARI"), nrow(mtcars), TRUE),
  player = sample(
  c("00-0033873", "00-0035228", "00-0036355", "00-0019596"),
  nrow(mtcars),
  TRUE
  )
)

ggplot(df, aes(x = mpg, y = disp)) +
  geom_point() +
  facet_wrap(vars(team)) +
  labs(
    title = tools::toTitleCase("These are random teams and data"),
    subtitle = "I just want to show how the nflplotR theme elements work",
    caption = "https://github.com/nflverse/nflseedR/raw/master/man/figures/caption.png"
  ) +
  theme_minimal() +
  theme(
    plot.title.position = "plot",
    plot.title = element_text(face = "bold"),
    axis.title = element_blank(),
    # load image from url in caption
    plot.caption = element_path(hjust = 1, size = 0.4)
  )
}
}
\seealso{
\code{\link[=geom_from_path]{geom_from_path()}} for more information.
}
