% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_from_path.R, R/ggpath-package.R
\docType{data}
\name{GeomFromPath}
\alias{GeomFromPath}
\alias{ggpath}
\alias{ggpath-package}
\title{ggpath: Robust Image Rendering Support for 'ggplot2'}
\format{
An object of class \code{GeomFromPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomFromPath
}
\description{
A 'ggplot2' extension that enables robust image grobs in panels and theme elements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrcaseb/ggpath}
  \item Report bugs at \url{https://github.com/mrcaseb/ggpath/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{datasets}
\keyword{internal}
