% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_from_path.R, R/geom_lines.R,
%   R/ggpath-package.R
\docType{data}
\name{GeomFromPath}
\alias{GeomFromPath}
\alias{GeomRefLines}
\alias{ggpath-package}
\title{ggpath: Robust Image Rendering Support for 'ggplot2'}
\format{
An object of class \code{GeomFromPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomRefLines} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomFromPath

GeomRefLines
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A 'ggplot2' extension that enables robust image grobs in panels and theme elements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrcaseb/ggpath}
  \item \url{https://mrcaseb.github.io/ggpath/}
  \item Report bugs at \url{https://github.com/mrcaseb/ggpath/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com} [copyright holder]

}
\keyword{datasets}
\keyword{internal}
